tests/requirements.txt
django==1.11
selenium==3.7.0
tox==2.9.1
django-nested-admin==3.0.20